Move type, enum to models
